% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census_divisions_2016}
\alias{census_divisions_2016}
\title{Census divisions (2016)}
\format{A data.frame with 91430 rows and 13 variables:
\describe{
  \item{long}{Longitude}
  \item{lat}{Latitude}
  \item{order}{Order of layers}
  \item{hole}{Polygon hole (TRUE or FALSE)}
  \item{piece}{Piece}
  \item{id}{Uniquely identifies a census division (composed of the 2-digit
            province/territory unique identifier followed by the 2-digit
            census division code).}
  \item{group}{Group}
  \item{census_division_name}{Census division name}
  \item{census_divison_type}{Census division type}
  \item{pr_alpha}{Province or territory 2-letter identifier}
  \item{pr_sgc_code}{Province Standard Geographical Classification
                           (SGC) code.}
  \item{pr_english}{Province name (English)}
  \item{pr_french}{Province name (French)}
}}
\source{
\url{https://www12.statcan.gc.ca/census-recensement/2011/geo/bound-limit/bound-limit-2016-eng.cfm}
(under Statistics Canada Open Licence \url{https://www.statcan.gc.ca/eng/reference/licence})
}
\usage{
census_divisions_2016
}
\description{
A data set with geographic information for Canadian census divisions
}
\keyword{datasets}
