% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{federal_riding_bins}
\alias{federal_riding_bins}
\title{Canadian federal riding bins (used for tile plots)}
\format{A data.frame with 944 rows and 8 variables:
\describe{
  \item{y}{y-axis of riding bins (corresponds to longitude)}
  \item{x}{x-axis of riding bins (corresponds to latitude)}
  \item{pr_alpha}{Province or territory 2-letter identifier}
  \item{representation_order}{Representation order. Specifies boundaries/number of seats for a given election
  (e.g. the 2015 election used the 2013 representation order, with 338 seats).}
  \item{pr_french}{Province or territory name (French).}
  \item{pr_english}{Province or territory name (English).}
  \item{pr_sgc_code}{Province Standard Geographical Classification
                           (SGC) code.}
  \item{riding_code}{Uniquely identifies a federal electoral district (composed of the
              2-digit province/territory unique identifier followed by the
              3-digit federal electoral district code).}
}}
\usage{
federal_riding_bins
}
\description{
A data set with coordinates for the \code{mapcan::riding_binplot()} function.
}
\keyword{datasets}
