% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{federal_riding_hexagons}
\alias{federal_riding_hexagons}
\title{Canadian federal riding hexagons (used for hexagonal tile plots)}
\format{A data.frame with 6629 rows and 15 variables:
\describe{
  \item{long}{y-axis of riding hexagons}
  \item{lat}{x-axis of riding hexagons}
  \item{order}{Order of layers}
  \item{hole}{Polygon hole (TRUE or FALSE)}
  \item{piece}{Piece}
  \item{group}{Group}
  \item{representation_order}{Representation order. Specifies boundaries/seats for a given election
  (e.g. the 2015 election used the 2013 representation order, with 338 seats).}
  \item{pr_french}{Province or territory name (French).}
  \item{pr_english}{Province or territory name (English).}
  \item{pr_sgc_code}{Province Standard Geographical Classification
                           (SGC) code.}
  \item{riding_code}{Uniquely identifies a federal electoral district (composed of the
              2-digit province/territory unique identifier followed by the
              3-digit federal electoral district code).}
}}
\usage{
federal_riding_hexagons
}
\description{
A data set with coordinates for the \code{mapcan::riding_binplot()} function.
}
\keyword{datasets}
