% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{province_pop_annual}
\alias{province_pop_annual}
\title{Annual provincial populations data frame dating back to 1971}
\format{A data.frame with 638 rows and 3 variables:
\describe{
  \item{province}{English name of province}
  \item{population}{Population of province}
  \item{year}{Year}
}}
\usage{
province_pop_annual
}
\description{
A data set with annual information on provincial and territorial populations dating back to 1971.
}
\keyword{datasets}
