% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{provinces_noterr_carto}
\alias{provinces_noterr_carto}
\title{Provinces and territories cartogram data (territories excluded)}
\format{A data.frame with 16797 rows and 11 variables:
\describe{
  \item{long}{Longitude}
  \item{lat}{Latitude}
  \item{order}{Order of layers}
  \item{hole}{Polygon hole (TRUE or FALSE)}
  \item{piece}{Piece}
  \item{pr_english}{Province or territory name (English).}
  \item{group}{Group}
  \item{population}{2016 Population of Province}
  \item{pr_alpha}{Province or territory 2-letter identifier}
  \item{pr_french}{Province or territory name (French).}
  \item{province_sgc_code}{Province Standard Geographical Classification
                           (SGC) code}
                           }}
\source{
\url{https://www12.statcan.gc.ca/census-recensement/2011/geo/bound-limit/bound-limit-2016-eng.cfm},
(under Statistics Canada Open Licence \url{https://www.statcan.gc.ca/eng/reference/licence})
}
\usage{
provinces_noterr_carto
}
\description{
A data set with geographic information for Canadian provinces and territories, boundary divisions
distorted by population size. Territories excluded.
}
\keyword{datasets}
