% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quebec_prov_ridings2018}
\alias{quebec_prov_ridings2018}
\title{Quebec provincial ridings geographic data}
\format{A data.frame with 23995 rows and 11 variables:
\describe{
  \item{long}{y-axis of riding hexagons}
  \item{lat}{x-axis of riding hexagons}
  \item{order}{Order of layers}
  \item{hole}{Polygon hole (TRUE or FALSE)}
  \item{piece}{Piece}
  \item{riding_code}{Uniquely identifies a provincial electoral district}
  \item{group}{Group}
  \item{riding_name}{Riding name (lowercase)}
  \item{riding_name}{Riding name (uppercase)}
  \item{centroid_long}{Longitude for riding centroids (useful for labeling)}
  \item{centroid_lat}{Latitude for riding centroids (useful for labeling)}
}}
\usage{
quebec_prov_ridings2018
}
\description{
A data set with geographic information for Quebec provincial ridings
}
\keyword{datasets}
