% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quebec_provincial_results}
\alias{quebec_provincial_results}
\title{Quebec provincial election results data}
\format{A data.frame with 125 rows and 6 variables:
\describe{
  \item{party}{Winning party of riding.}
  \item{vote_share}{Percentage of vote won by winning candidate.}
  \item{riding_code}{Uniquely identifies a provincial electoral district}
  \item{riding_name}{Riding name (lowercase)}
  \item{riding_name}{Riding name (uppercase)}
}}
\usage{
quebec_provincial_results
}
\description{
A data set with information on 2018 Quebec provincial election results
}
\keyword{datasets}
