% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quebec_riding_bins}
\alias{quebec_riding_bins}
\title{Quebec provincial riding bins (used for tile plots)}
\format{A data.frame with 125 rows and 6 variables:
\describe{
  \item{y}{y-axis of riding bins (corresponds to longitude)}
  \item{x}{x-axis of riding bins (corresponds to latitude)}
  \item{riding_code}{Riding code}
  \item{region}{Region}
  \item{riding_simplified}{Simplified riding name}
  \item{riding_name}{Riding name}
}}
\usage{
quebec_riding_bins
}
\description{
A data set with coordinates for the \code{mapcan::riding_binplot()} function.
}
\keyword{datasets}
