% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quebec_riding_hexagons}
\alias{quebec_riding_hexagons}
\title{Quebec provincial riding hexagons (used for hexagonal tile plots)}
\format{A data.frame with 6629 rows and 15 variables:
\describe{
  \item{long}{y-axis of riding hexagons}
  \item{lat}{x-axis of riding hexagons}
  \item{order}{Order of layers}
  \item{hole}{Polygon hole (TRUE or FALSE)}
  \item{piece}{Piece}
  \item{group}{Group}
  \item{y}{y-axis of riding hexagon center}
  \item{x}{x-axis of riding hexagon center}
  \item{region}{Region}
  \item{riding_simplified}{Simplified riding name}
  \item{riding_name}{Riding name}
  \item{riding_code}{Riding code}
}}
\usage{
quebec_riding_hexagons
}
\description{
A data set with coordinates for the \code{mapcan::riding_binplot()} function.
}
\keyword{datasets}
