% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{riding_info}
\alias{riding_info}
\title{Canadian federal riding population information}
\format{A data.frame with 37111 rows and 12 variables:
\describe{
  \item{party}{Winning party in riding}
  \item{riding_code}{Riding code}
  \item{population_2011}{Population of riding in 2011}
  \item{population_2016}{Population of riding in 2016}
}}
\source{
\url{http://www.elections.ca/content.aspx?section=ele&dir=pas&document=index&lang=e},
(under Open Government Licence \url{https://open.canada.ca/en/open-government-licence-canada})
}
\usage{
riding_info
}
\description{
A data set with information on Canadian federal election results, dating back to 1997
}
\keyword{datasets}
