% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mapcan.R
\name{theme_mapcan}
\alias{theme_mapcan}
\title{Mapcan theme}
\usage{
theme_mapcan(legend_position = "bottom", base_size = 12,
  base_family = "")
}
\description{
A ggplot theme that removes unecessary components of map plots. Builds on \code{theme_bw()}.
}
