% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{federal_ridings}
\alias{federal_ridings}
\title{Federal ridings}
\format{A data.frame with 46830 rows and 15 variables:
\describe{
  \item{long}{Longitude}
  \item{lat}{Latitude}
  \item{order}{Order of layers}
  \item{hole}{Polygon hole (TRUE or FALSE)}
  \item{piece}{Piece}
  \item{riding_code}{Uniquely identifies a federal electoral district (composed of the
              2-digit province/territory unique identifier followed by the
              3-digit federal electoral district code).}
  \item{group}{Group}
  \item{riding_name_english}{Federal electoral district name in English.}
  \item{riding_name_french}{Federal electoral district name in French.}
  \item{province_sgc_code}{Province Standard Geographical Classification
                           (SGC) code}
  \item{pr_english}{Province name (English)}
  \item{pr_french}{Province name (French)}
  \item{pr_alpha}{Province or territory 2-letter identifier}
}}
\source{
\url{https://www12.statcan.gc.ca/census-recensement/2011/geo/bound-limit/bound-limit-2016-eng.cfm},
(under Statistics Canada Open Licence \url{https://www.statcan.gc.ca/eng/reference/licence})
}
\usage{
federal_ridings
}
\description{
A data set with geographic information for Canadian federal
ridings (2013 representation order)
}
\keyword{datasets}
