% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mapcan.R
\name{theme_mapcan}
\alias{theme_mapcan}
\title{Mapcan theme}
\usage{
theme_mapcan(legend_position = "bottom", base_size = 12,
  base_family = "")
}
\arguments{
\item{legend_position}{Position of legend, default is "bottom"}

\item{base_size}{Base font size (default is 12)}

\item{base_family}{Base font family}
}
\description{
A ggplot theme that removes unnecessary components of map plots. Builds on \code{theme_bw()}.
}
